Update `Attribute`s for semconv
Update usage of `Attribute`s to be in compliance with semantic
conventions version 1.23.1.
